refactor: extract shared AgentContentMapper in camunda- and rdbms-exporter
Move the content-block mapping logic out of AgentHistoryHandler and
AgentHistoryExportHandler into standalone AgentContentMapper utilities
so it can be reused for the agent instance system prompt. Behavior is
unchanged.
